requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1
scikit-image>=0.21.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
flask>=3.0.0
flask-cors>=4.0.0
cloudinary>=1.36.0
scipy>=1.11.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
